Rename package and application class.
webapiskeleton => musicapi
SkeletonApplication => MusicApplication
