+++ gradle.properties
@@ -1,8 +1,8 @@
 # properties used in build.gradle
 
 # application properties
-mainClass=edu.oregonstate.mist.webapiskeleton.SkeletonApplication
-archivesBaseName=web-api-skeleton
+mainClass=edu.oregonstate.mist.musicapi.MusicApplication
+archivesBaseName=music-collection-api
 
 # library properties
 groovyVersion=2.4.3
